build & fs: build fixes and refined fs
