--- app/components/footer/footer.module.css
@@ -1,76 +1,0 @@
-/* Copyright */
-.copyright{
-  box-sizing: border-box;
-
-/* Auto layout */
-display: flex;
-flex-direction: row;
-justify-content: space-between;
-align-items: center;
-padding: 32px 0px;
-gap: 200px;
-margin-top: 4rem;
-width: auto;
-height: 108px;
-
-/* Primary/Black */
-background: #282A2C;
-border-top: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-/* Social Icon */
-.social{
-  display: flex;
-flex-direction: row;
-align-items: flex-start;
-padding: 0px;
-gap: 8px;
-
-margin: 0 auto;
-width: 200px;
-height: 44px;
-margin-left: 0px;
-}
-
-
-/* Social Media */
-.socialmedia{display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  padding: 12px;
-  gap: 10px;
-  
-  width: 44px;
-  height: 44px;
-  
-  background: rgba(255, 255, 255, 0.1);
-  border-radius: 5px;
-}
-/* Auto layout */
-
-
-
-/* Social */
-.icondiv{
-  width: 20px;
-height: 20px;
-}
-
-/* © 2021 - DummyCompany */
-.trademark{
-  margin: 0 auto;
-  margin-right: 0px;
-width: 191px;
-height: 24px;
-
-font-family: 'Satoshi';
-font-style: normal;
-font-weight: 400;
-font-size: 16px;
-line-height: 24px;
-/* identical to box height, or 150% */
-text-align: center;
-
-/* Primary/White */
-color: #FFFFFF;
-}
