--- app/components/hero/hero.module.css
@@ -1,179 +1,0 @@
-.herosection{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 0px;
-    gap: 40px;
-    position: absolute;
-    width: 493px;
-    height: 334px;
-    left: 34% ;
-    top: 196px;
-}
-
-.herosectiontext{
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 0px;
-    gap: 20px;
-
-    width: 493px;
-    height: 238px;
-}
-
-
-.herotitle{
-    width: 493px;
-    height: 128px;
-
-    font-family: 'Satoshi';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 56px;
-    line-height: 64px;
-    /* or 114% */
-    text-align: center;
-    letter-spacing: -1px;
-
-    color: #FFFFFF;
-}
-
-.herotext{
-    width: 490px;
-    height: 90px;
-
-    font-family: 'Satoshi';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 21px;
-    line-height: 30px;
-    /* or 143% */
-    text-align: center;
-
-    color: rgba(255, 255, 255, 0.8);
-
-}
-
-.herobtn{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 13px 24px;
-    gap: 10px;
-
-    width: 160px;
-    height: 56px;
-
-    background: #E43333;
-    border-radius: 4px;
-}
-
-.herobtntext{
-    width: 80px;
-    height: 24px;
-
-    font-family: 'Satoshi';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 16px;
-    line-height: 150%;
-    /* identical to box height, or 24px */
-    text-align: center;
-    letter-spacing: -0.02em;
-
-    color: #FFFFFF;
-}
-
-
-.herofooter{
-width: auto;
-height: 270px;
-/* margin-top: 600px; */
-position: relative;
-display: flex;
-flex-direction: row;
-justify-content: center;
-align-items: center;
-
-/* Greyscale/Grey 10 */
-background: #F7F8FB;
-}
-
-
-
-.herofootertext{
-display: flex;
-flex-direction: row;
-align-items: flex-start;
-padding: 0px;
-gap: 24px;
-
-position: absolute;
-width: auto;
-height: 70px;
-left: auto;
-top: 100px;
-}
-/* Auto layout */
-
-
-
-
-/* Fun-Fact */
-.fact{
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 0px;
-    gap: 20px;
-    
-    width: 266px;
-    height: 70px;
-}
-/* Auto layout */
-
-
-/* 17+ */
-.no{
-    width: 266px;
-    height: 34px;
-
-    font-family: 'Bebas Neue';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 48px;
-    line-height: 56px;
-    /* or 117% */
-    /* leading-trim and text-edge are draft CSS properties.
-    
-    Read more: https://drafts.csswg.org/css-inline-3/#leading-trim
-    */
-    text-align: center;
-    letter-spacing: -0.015em;
-    
-    /* Primary/Red */
-    color: #E43333;
-}
-
-
-/* Awards Rewarded */
-.title{
-    width: 266px;
-    height: 16px;
-    
-    font-family: 'Satoshi';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 100%;
-    /* identical to box height, or 16px */
-    text-align: center;
-    letter-spacing: 0.01em;
-    text-transform: uppercase;
-    
-    /* Primary/Black */
-    color: #282A2C;
-}
