--- app/components/review/review.tsx
@@ -1,63 +1,0 @@
-import React from 'react'
-
-const Review = () => {
-  return (
-    <div className='relative w-auto h-[708px] flex flex-col mt-16'>
-        <div className='top-20 left-20 width-[1136px] h-[508px]'>
-            <div className='width-auto h-[116px] flex flex-col justify-between'>
-                <p className='text-4xl font-semibold ml-[140px]'>What Client Says</p>
-                <p className='text-[#75777A] ml-[140px] w-[547px]'>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat aute irure sint amet</p>
-            </div>
-            <div className='flex flex-row justify-between w-[1136px] h-[320px] gap-10 ml-[140px] mt-20'>
-                <div className='w-[380px] h-[320px] rounded-xl bg-[#F3F5F7]'>
-                    <div className='w-[360px] h-[240px] flex flex-col mt-4 ml-5'>
-                        <div className='flex flex-row ml-4 mt-4'>
-                            <img src="pp1.svg" alt="profile pic"/>
-                            <div className='flex flex-col ml-5'>
-                                <div>Annete Black</div>
-                                <div className='flex flex-row'>
-                                <div className='text-[#75777A]'>Chief Chairman of</div>
-                                <div className='font-bold ml-1'>Netflix</div>
-                                </div>
-                            </div>
-                        </div>
-                        <p className='mt-6 mr-4 ml-4'>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat aute irure sint amet occaecat cupidatat non proident ea commodo consequat aute irure sint amet occaecat cupidatat non proident ea commodo </p>
-                    </div>
-                </div>
-                <div className='w-[380px] h-[320px] rounded-xl bg-[#F3F5F7] ml-8'>
-                    <div className='w-[360px] h-[240px] flex flex-col mt-4 ml-5'>
-                        <div className='flex flex-row ml-4 mt-4'>
-                            <img src="pp1.svg" alt="profile pic"/>
-                            <div className='flex flex-col ml-5'>
-                                <div>Annete Black</div>
-                                <div className='flex flex-row'>
-                                <div className='text-[#75777A]'>Chief Chairman of</div>
-                                <div className='font-bold ml-1'>Netflix</div>
-                                </div>
-                            </div>
-                        </div>
-                        <p className='mt-6 mr-4 ml-4'>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat aute irure sint amet occaecat cupidatat non proident ea commodo consequat aute irure sint amet occaecat cupidatat non proident ea commodo </p>
-                    </div>
-                </div>
-                <div className='w-[380px] h-[320px] rounded-xl bg-[#F3F5F7] ml-8'>
-                    <div className='w-[360px] h-[240px] flex flex-col mt-4 ml-5'>
-                        <div className='flex flex-row ml-4 mt-4'>
-                            <img src="pp1.svg" alt="profile pic"/>
-                            <div className='flex flex-col ml-5'>
-                                <div>Annete Black</div>
-                                <div className='flex flex-row'>
-                                <div className='text-[#75777A]'>Chief Chairman of</div>
-                                <div className='font-bold ml-1'>Netflix</div>
-                                </div>
-                            </div>
-                        </div>
-                        <p className='mt-6 mr-4 ml-4'>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat aute irure sint amet occaecat cupidatat non proident ea commodo consequat aute irure sint amet occaecat cupidatat non proident ea commodo </p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-  )
-}
-
-export default Review
