+++ app/page1.json
@@ -10,7 +10,7 @@
         "i":"/product/p11.svg"
     },
     {
-        "name":"PORTABLE FIRE EXTINGUISHERS (GAS CARTRIDGE) POWDER TYPE , (IS:15683)",
+        "n":"PORTABLE FIRE EXTINGUISHERS (GAS CARTRIDGE) POWDER TYPE , (IS:15683)",
         "d":"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
         "f":["Fire Extinguisher (Gas Cartridge type) Capacity 04 Kg duly ISI marked conforming  to IS: 15683:2006, with IS 14609, ISI marked ABC powder; Complete with IS 4947 ISI marked Gas Cartridge with squeeze grip release valve, discharge, pipe locking arrangement, pressure gauge, bracket. Operating Temperature: - 10ºC to + 55ºC. GETECH Make",
